Added signIn hook and refactored file structure
