+++ src/components/dynamic_searchbar.jsx
