+++ src/components/signUp.jsx
@@ -1,11 +1,12 @@
 import React, {Component} from "react";
-import "./signUp.css";
-import {signUpFirebase} from "./signUpHooks";
+import "../css/signUp.css";
+import {signUpFirebase} from "../hooks/signUpHooks";
+
 class SignUp extends Component {
     doSignUp()
     {
         console.log(this.state.userEmail.indexOf("mail.gvsu.edu"));
-        if(this.state.userEmail.indexOf("mail.gvsu.edu") != -1)
+        if(this.state.userEmail.indexOf("mail.gvsu.edu") !== -1)
         {
             signUpFirebase(this.state.userEmail, this.state.userPassword);
         }
