+++ src/components/simple_searchbar.jsx
