+++ src/css/signUp.css
@@ -30,27 +30,27 @@
     text-transform: uppercase;
     letter-spacing: .15rem;
     border: 0;
-  }
-  
-  .btn-primary {
+}
+
+.btn-primary {
     background-color: #64a19d;
     margin: 1.25rem;
-  }
-  
-  .btn-primary:hover {
+}
+
+.btn-primary:hover {
     background-color: #4f837f;
-  }
-  
-  .btn-primary:focus {
+}
+
+.btn-primary:focus {
     background-color: #4f837f;
     color: white;
-  }
-  
-  .btn-primary:active {
-    background-color: #467370 !important;
-  }
+}
 
-  .btn-google {
+.btn-primary:active {
+    background-color: #467370 !important;
+}
+
+.btn-google {
     box-shadow: 0 0.1875rem 0.1875rem 0 rgba(0, 0, 0, 0.1) !important;
     padding: 1.25rem 2rem;
     font-family: 'Varela Round';
@@ -60,4 +60,4 @@
     border: 0;
     background-color: #ddeceb;
     margin-bottom: 1.25rem;
-  }
+}
